Internally split `EvaluationContext` into its own file
